@@ -4,15 +4,15 @@ apply plugin: 'kotlin-kapt'
 apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         multiDexEnabled true
         applicationId "com.tompee.convoy"
         minSdkVersion 17
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables {
             useSupportLibrary = true
         }
@@ -37,18 +37,16 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
-    def support_library_version = "27.1.1"
-    implementation "com.android.support:cardview-v7:$support_library_version"
-    implementation "com.android.support:animated-vector-drawable:$support_library_version"
-    implementation "com.android.support:customtabs:$support_library_version"
-    implementation "com.android.support:appcompat-v7:$support_library_version"
-    implementation "com.android.support:design:$support_library_version"
+    def support_library_version = "1.0.0"
+    implementation "androidx.cardview:cardview:$support_library_version"
+    implementation "androidx.appcompat:appcompat:$support_library_version"
+    implementation "com.google.android.material:material:$support_library_version"
 
     def constraint_layout_version = "1.1.3"
-    implementation "com.android.support.constraint:constraint-layout:$constraint_layout_version"
+    implementation "androidx.constraintlayout:constraintlayout:$constraint_layout_version"
 
-    def multidex_version = "1.0.3"
-    implementation "com.android.support:multidex:$multidex_version"
+    def multidex_version = "2.0.0"
+    implementation "androidx.multidex:multidex:$multidex_version"
 
     def facebook_version = "4.30.0"
     implementation "com.facebook.android:facebook-login:$facebook_version"
@@ -71,7 +69,7 @@ dependencies {
     def firebase_auth = "16.0.3"
     implementation "com.google.firebase:firebase-auth:$firebase_auth"
 
-    def firebase_storage = "16.0.1"
+    def firebase_storage = "16.0.2"
     implementation "com.google.firebase:firebase-storage:$firebase_storage"
 
     def firebase_messaging = "17.3.2"
@@ -92,7 +90,7 @@ dependencies {
     def rx_binding_version = "2.1.1"
     implementation "com.jakewharton.rxbinding2:rxbinding-kotlin:$rx_binding_version"
 
-    def dagger_version = "2.14.1"
+    def dagger_version = "2.15"
     implementation "com.google.dagger:dagger:$dagger_version"
     implementation "com.google.dagger:dagger-android:$dagger_version"
     implementation "com.google.dagger:dagger-android-support:$dagger_version"
