@@ -1,8 +1,8 @@
 package com.tompee.convoy.base
 
 import android.os.Bundle
-import android.support.annotation.LayoutRes
-import android.support.v7.app.AppCompatActivity
+import androidx.annotation.LayoutRes
+import androidx.appcompat.app.AppCompatActivity
 import android.view.MenuItem
 
 abstract class BaseActivity : AppCompatActivity() {
