@@ -3,7 +3,7 @@ package com.tompee.convoy.feature.login
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
-import android.support.v4.view.ViewPager
+import androidx.viewpager.widget.ViewPager
 import android.view.View
 import com.tompee.convoy.ConvoyApplication
 import com.tompee.convoy.R
