@@ -1,7 +1,7 @@
 package com.tompee.convoy.feature.login
 
-import android.support.v4.app.FragmentManager
-import android.support.v4.app.FragmentStatePagerAdapter
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentStatePagerAdapter
 import com.tompee.convoy.feature.login.page.LoginFragment
 
 class LoginPagerAdapter(fragmentManager: FragmentManager,
