@@ -1,11 +1,11 @@
 package com.tompee.convoy.feature.friend.adapter
 
-import android.support.v7.widget.RecyclerView
 import android.util.SparseArray
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
 import com.tompee.convoy.R
 import java.util.*
 
