@@ -2,7 +2,7 @@ package com.tompee.convoy.feature.profilesetup
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.design.widget.Snackbar
+import com.google.android.material.snackbar.Snackbar
 import com.jakewharton.rxbinding2.view.RxView
 import com.jakewharton.rxbinding2.widget.RxTextView
 import com.tompee.convoy.ConvoyApplication
